requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
frozenlist==1.3.1
future==0.18.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
junos-eznc==2.6.5
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
napalm==4.0.0
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.1.2
netutils==1.2.0
ntc-templates==3.1.0
paramiko==2.11.0
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
pyeapi==0.8.4
pyee==9.0.4
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PyYAML==6.0
requests==2.28.1
scp==0.14.4
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
tenacity==8.1.0
textfsm==1.1.2
tomli==2.0.1
transitions==0.9.0
ttp==0.9.1
ttp-templates==0.3.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yamlordereddictloader==0.4.0
yarl==1.8.1
